requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
<<<<<<< HEAD
=======
celery==5.0.2
>>>>>>> 76139b7ca3cf11128f150c30d131b88c8d03e879
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
<<<<<<< HEAD
dj-database-url==0.5.0
Django==3.1.3
django-bulk-update==2.2.0
=======
Django==3.1.3
django-bulk-update==2.2.0
django-celery-beat==2.1.0
>>>>>>> 76139b7ca3cf11128f150c30d131b88c8d03e879
django-cors-headers==3.5.0
django-timezone-field==4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
kombu==5.0.2
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
